moved linalg one directory up
