+++ heat/core/__init__.py
@@ -6,7 +6,6 @@ from .exponential import *
 from .factories import *
 from .indexing import *
 from .io import *
-from .linalg import *
 from .logical import *
 from .manipulations import *
 from .memory import *
@@ -16,7 +15,6 @@ from .relational import *
 from .rounding import *
 from .statistics import *
 from .dndarray import *
-from .tiling import *
 from .trigonometrics import *
 from .types import *
 from .types import finfo, iinfo
