+++ heat/linalg/qr.py
@@ -1,10 +1,9 @@
 import collections
 import torch
 
-from .. import dndarray
-from .. import logical
-from .. import factories
-from .. import tiling
+from heat.core import dndarray
+from heat.core import factories
+from heat.linalg import tiling
 
 __all__ = ["qr"]
